feature: add arena serve logs and refactor some codes
