+++ pkg/operators/tf-operator/apis/tensorflow/v1/defaults_test.go
@@ -18,10 +18,10 @@ import (
 	"reflect"
 	"testing"
 
-	"k8s.io/api/core/v1"
+	v1 "k8s.io/api/core/v1"
 
-	common "github.com/kubeflow/arena/pkg/tf-operator/apis/common/v1"
-	"github.com/kubeflow/arena/pkg/tf-operator/util"
+	common "github.com/kubeflow/arena/pkg/operators/tf-operator/apis/common/v1"
+	"github.com/kubeflow/arena/pkg/operators/tf-operator/util"
 )
 
 const (
