+++ pkg/operators/tf-operator/apis/tensorflow/validation/validation_test.go
@@ -18,10 +18,10 @@ import (
 	"testing"
 
 	"github.com/golang/protobuf/proto"
-	commonv1 "github.com/kubeflow/arena/pkg/tf-operator/apis/common/v1"
-	commonv1beta2 "github.com/kubeflow/arena/pkg/tf-operator/apis/common/v1beta2"
-	tfv1 "github.com/kubeflow/arena/pkg/tf-operator/apis/tensorflow/v1"
-	tfv1beta2 "github.com/kubeflow/arena/pkg/tf-operator/apis/tensorflow/v1beta2"
+	commonv1 "github.com/kubeflow/arena/pkg/operators/tf-operator/apis/common/v1"
+	commonv1beta2 "github.com/kubeflow/arena/pkg/operators/tf-operator/apis/common/v1beta2"
+	tfv1 "github.com/kubeflow/arena/pkg/operators/tf-operator/apis/tensorflow/v1"
+	tfv1beta2 "github.com/kubeflow/arena/pkg/operators/tf-operator/apis/tensorflow/v1beta2"
 
 	"k8s.io/api/core/v1"
 )
